split nosql engine
split nosql into 3 files for easier management
